Fail module-deps.pl on more errors
* Fail on the error code returned from make
* Collect dependency errors in the tree and fail and print them

The "perldoc -f open" has an example in the "Opening a filehandle
into a command" section, which uses waitpid in the parent. Since
we parse the pipe output, there seems to be no need for waitpid,
because we end when the pipe is closed by the writer.
waitpid always returns -1 at this point and since there aren't
any zombie processes, it seems to be fine to ignore.
Additionally strace shows, clone, execve and waitpid calls, so
the explicit exit in the example seems bogus, if the process
is actually overwritten.
And the pipe is just open in the parent process anyway.

Change-Id: I2e77aa33a0eaa5d6d3e06904bb0af53f4b66ef91
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126789
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>