tdf#120896 sc: commit change after PopupSpelling
Although PopupSpelling visually has changed the screen,
the underlying document still contained the
old cached string or spelling status.
Thus spell-check still thought it was spelled wrong,
and right-clicking got very confused.

Steps to reproduce:
1.) type aaa and hit enter.  (spelling error noted)
2.) right-click on "aaa" and chose various options.
2a.) --- replace with a suggestion
2b.) --- ignore word
2c.) --- add to dictionary
2d.) --- change language
2d.) --- autocorrect
(note we are still in editing mode)
3.) right-click again on the new or approved word
(no pop-up at all.)

The fix is to stop editing, so the changes are
applied to the various caches.

Steps that should NOT stop editing the cell
2e.) --- press ESC (or anything that cancels the pop-up)
2f.) --- choose Autocorrect Options

Change-Id: I9965ede6f6a5c31a811ff89abb687497306d6d1e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126802
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins