tdf#135061 sw_redlinehide: create frames following hidden table
When pasting, the entire document body is covered by a delete redline.

The insert position is in the last node, following a table; table nodes
do hot have any connection to frames if they are hidden by redlines,
so MakeFrames() won't create frames for the inserted nodes.

Refactor SwNodes::FindPrvNxtFrameNode() so that it can detect that table
nodes are hidden in the current layout and continue to search in this
case.

For multiple layouts, MakeFrames() may need to use different nodes per
layout, as the "nearest" one in one layout may be hidden in another.

Change-Id: I3bb2c861c861438ac2695ab49dd91dc2bde87db4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127272
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>