tdf#138743 sw_redlinehide: delete FlyFrames when hiding tracked changes
Cannot remember why but when RemoveHiddenObjsOfNode() was added in
commit 2b16182c7282b4ad119d460c70c0e3272f1f3172 only as-char draw shapes
were removed from the draw page, but as-char fly frames remained.

For these only a barely visible border is painted sometimes, but the
problem is they remain as a mouse click target, not sure how i didn't
notice this earlier.

Somehow image painting is wonky on current master, works fine on 7.1.

Change-Id: Ie4e20f6a399daadda1badf155ba510445394f5e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127366
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>