Further separate saxparser run from cmp'n'touch
At least on Windows the calls to saxparser almost feel as slow
as the compilation. So this adds an intermediate file just for
the cmp calls, costing just a bit more filesystem space.

This also reverts the gb_MKTEMP usage introduced in commit
0e14a36a408c7ea07d0b6d78a90b10e3075c03bd. Using a fixed name in
WORKDIR is easier to debug eventually, then some random mktemp
generated file, and is faster anyway.

Change-Id: Id4979b75b881e916c209ff0ff8e123536294e93b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127841
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>