tdf#129578 rtfimport: CJK char properties are not CS
This fixes a LO 6.4/6.3.0.1 regression caused by
commit ef21a736197bfa6725e836dc1c94c7c746eba477
which was focused on export, and not import.

The code originally didn't have any
mention of DBCH, so it used the default.
However, in the 6.4 change, it was wrongly tied together
with the Complex Script indicator.

So revert that part of the change.
Mark fixed the same problem in 7.1 for font size.

Change-Id: I1c87ae856de93ebc693135585574f2eb144b214b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128113
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>