tdf#146605 sw: try to fix SwSectionFormat notifications
SwSectionFormat no longer notifies via its SwModify to SwClient
listeners, but only via SvtBroadcaster, so SwSection and SwSectionFrame
need to be SvtListeners.

This can be seen when changing the footnote-at-end setting in the
section dialog, if the SwSection doesn't get the event the footnote
numbers aren't adapted to a new restart value, if the SwSectionFrame
doesn't get the event the footnote container isn't moved.

This is of course not pretty but hopefully it will be refactored further
in the future.

(regression from commit b1490bb8beea265d70b21a66a5c35cceaa8040dd)

Change-Id: I93c616e209a01fc5d4c8eba701a6ddce5d5b432e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128257
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>