tdf#145321 Crash scrolling DOCX to bottom
this is a consequence of
    commit dc6c180c14b18c6ff1a5b079961acdbd42277efc
    use std::vector for fetching DX array data
which made pre-existing bugs easier to see.

This crash made apparent that we have bad data ending up in
SwDrawTextInfo. So I added some asserts there to catch that.

However, that simply made apparent that there are bug(s) at
a higher level that I have no idea how to to fix.

So at the primary problem site in
SwTextCursor::GetModelPositionForViewPoint I clamp
the values to sane ones.

Change-Id: Ic74f6944932bbfc22e8cf9addf9e7511755b18be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128497
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>