tdf#146893 Rework Manage Changes dialog
This patch reworks the "Manage Changes" dialog (In Writer, Edit > Track Changes > Manage) so that the buttons "Accept", "Accept all", "Reject", "Reject all" and "Undo" only appear in the List tab, since they're not needed in the Filter tab.

A help button is also added and HID constants are defined for both tabs.

Change-Id: I12da7811635dfc142d5d0a66d3275d3d170692e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128563
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>