v2 tdf#134951 docxexport: don't lose postponed comment
This patch basically reverts and re-writes the original fix.

Prior to mstahl's 7.4 commit 73c40b4093a537d46e5bfeb1445af330df657873
   tdf#143039 tdf#143106 sw: DOCX export: don't export duplicate flys
ooxmlexport5's comment would only last for one or two round-trips,
and then be lost after that. It kept moving around into things
like the fly itself, or the footnote.
With mstahl's commit, it was lost immediately (same as Word 2016),
and so the ooxmlexport5 unit test that broke was commented out.

Change-Id: Ia7e881143f1243a3ab81bb94c02b57189f3730f0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128703
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Justin Luth <jluth@mail.com>