SwNavigator: move content type init into SwContentType constructor
With commit cda8579cb41cd37c6acd7026675fa8ee7282314d the Init function
is no longer needed to help determine if content has changed. It's
other duty of setting wheather content type can be edited or deleted or
both, and if it has attachment for URL are moved to the constructor.

Change-Id: Ie6cf2273baca8e84e2b0d4757794e91f0a1af2dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126799
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>