tdf#111733 TextCharacterSpacingControl::GrabFocus kerning spinbutton
When the kerning menu opens, it defaulted to very tight,
regardless of what the current setting was.
However, the current value can easily be connected to the
menu options, so lets do that.

This depends on increasing the spinbutton range beyond -2.0,
done via LO 7.4 commit 691331640b3187b49184884b56ad8f12a69af938

Change-Id: I0be0956bf1cc3604faecc691aeac70a5bbba807b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128909
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>