Fix gtk3_kde5 build when qt5/kf5 VCL plugins are not built
Since

    commit f2b3e136da1528d89c3bd9d30726ac014e9e2cbd
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Sep 28 08:36:44 2021 +0200

        Unify how Qt/KF build flags/libs are set

the qt5/kf5 externals are used to set includes and
compiler/linker flags for the gtk3_kde5 VCL plugin
as well.

However, the relevant link targets were only actually
defined as needed when building the qt5 and kf5 VCL plugins
was enabled as well.
Adapt that so that building the gtk3_kde5 VCL plugin
but not the qt5 and kf5 ones also works again
(e.g. '--enable-gtk3-kde5 --disable-qt5 --disable-kf5').

Change-Id: I65b75d7c2ccab71eade01aa2529fb45174056701
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129022
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>