tdf#135843 Implement inside horizontal vertical borders.
wholeTbl as TableStylePart should be handled in different way. Before
left border of the whole table was handling like all cells left
border but it should be left border of the first column.

insideV and and insideH properties are imported but never handled. I
added the inside vertical and horizontal borders handling.

Change-Id: I5aea5cbefc746db637eac6c1438fa70a7d741bc0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128971
Tested-by: Jenkins
Reviewed-by: Gülşah Köse <gulsah.kose@collabora.com>