ScriptForge - (SF_Platform) refine locales
So far, only SF_Platform.Locale was supported, returning
the locale of the operating system.

Now, 3 different locales may be considered:

- SystemLocale: identical to Locale (still supported)
- OfficeLocale: the locale associated with the LibreOffice
  installation
- FormatLocale: the locale associated with how numbers and
  dates are formatted.

Of course, in many cases all locales are identical. Anyway
a distinction can be made.

All these properties are read-only.

Those properties are supported in Basic and Python evenly.

Change-Id: I0792d49a052b1441a30ac231d61eda997dc4216a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129297
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>