Remove COMPILER_PLUGINS_CXX from distro-configs/Jenkins/linux_clang_dbgutil_64
It had originally been added with a8fa6a3e2612d471cba256cd64db218ec533def5
"Remove CXXFLAGS_CXX11 from Clang plugin compilation", so "if
COMPILER_PLUGINS_CXX is not set, simply default it to g++ instead of trying to
construct an acceptable CLANGCXX value from CXX (which would be Clang).  (The
problem with using Clang without CXXFLAGS_CXX11 is that Clang, unlike GCC,
typically defaults to C++03, but building compilerplugins requires C++11 at
least.  That would cause e.g. the Gerrit/Jenkins linux_clang_dbgutil_64 builds
to fail---but which also needs COMPILER_PLUGINS_CXX to be explicitly set to 'g++
-std=c++11' as GCC on those machines is still 4.8.5 defaulting to C++03."  But
that should no longer be an issue with contemporary Clang, which defaults to >=
C++11 for quite a while now.

On the other hand, when trying to update the Clang used by
<https://ci.libreoffice.org/job/gerrit_linux_clang_dbgutil/> from 5.0.2 to
12.0.1, and adding

> export COMPILER_PLUGINS_CXX="ccache $LODE_HOME"/opt_private/gcc-7.3.0/bin/g++

to
<https://git.libreoffice.org/lode/+/refs/heads/master/bin/linux_clang_dbgutil_64.env>
(where this setting arguably belongs, rather than in
distro-configs/Jenkins/linux_clang_dbgutil_64, anyway), which is needed
because that version of Clang (and thus loplugin built against it)
cannot be built with the baseline CentOS 7 GCC 4.8.5, that setting of
COMPILER_PLUGINS_CXX got overriden by the one in
distro-configs/Jenkins/linux_clang_dbgutil_64, and configure failed due to

> configure:21069: checking clang/Basic/SourceLocation.h usability
> configure:21069: ccache g++ -std=c++11 -c -I/home/tdf/sberg/lode/opt_private/clang-llvmorg-12.0.1/include -std=c++14   -fno-exceptions -fno-rtti -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -I/home/\
> tdf/sberg/lode/packages/llvm-llvmorg-12.0.1.src/clang/include -I/home/tdf/sberg/lode/opt_private/clang-llvmorg-12.0.1/include -std=c++14   -fno-exceptions -fno-rtti -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_L\
> IMIT_MACROS -I/home/tdf/sberg/lode/packages/llvm-llvmorg-12.0.1.src/clang/include conftest.cpp >&5
> g++: error: unrecognized command line option '-std=c++14'
> g++: error: unrecognized command line option '-std=c++14'
> configure:21069: $? = 1

Change-Id: Ic33b116090f648ef645febb4fbb28ceb6a2a7cae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129692
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>