tdf#113785 sc FillAutoSimple: also optimize in negative direction
By only coding the optimization for the downward autofill,
it totally broke an upward autofill.
So this patch simply fixes the LO 4.3ish regression
from commit 293f59e26531f6dccc2c6dd4fe3c39c5ccd4ee1c.

This fix means that the non-hidden cells get the auto-value
instead of being erased.

However, the hidden cells are still erased
(which is the topic of tdf#119957).

I'm not going to bother with a unit test for this
portion. If I can solve bug 119957 as well,
then the unit test for that can cover both situations.

Change-Id: If6320ccf87fa8893ca6766c265b7760cc46ed7d8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129723
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>