ScriptForge - (SF_Utils) review optional argument of _Repr()
As from LO 7.3, optional missing arguments must not be used
in assignments and comparisons.
The Basic interpreter tolerated it before 7.3.

Probable commit having changed Basic behaviour:
    https://gerrit.libreoffice.org/c/core/+/121794

A unique case of this singularity was found in the
    SF_Utils._Repr()
method.

The bad consequence was that the commonly used
    SF_Exception.DebugPrint()
method calls above method. This broke in fact
the whole ScriptForge library.

Change-Id: Iba8792b92c73c992429de4382bd8a3dc4b58a94f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129728
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>